requirements.txt
Flask
Flask-Cors
requests
python-dotenv
firebase-admin
qrcode
pillow
weasyprint
gunicorn
